Near finished app
Music plays on startup (can be paused with music button top right)
can scroll through collection view containing 718 collection view cells
for the 718 pokemon, shows image and label for each.
Search bar can be used to search for a specific pokemon.
If a pokemon is selected it will segue into second view controller,
download pokemon data with an API call to pokeapi.co using the
Alamofire pod. when data is finished downloading the variables are set
with getters.
once variables are set the UI is updated to display information for
that pokemon (height, weight, attack, etc.)