Completed a feature, and added a new one
Completed the details screen to update UI with the next evolution’s
image if there is one, as well as the next evolutions pokemon name and
the level with which it evolves at.

Also added a button to enable or disable shiny sprites. When enabled,
each pokemon sprite will have its colours changed to its shiny
variation. This affects the main screens collection view as well as the
details view controller.